requirements.txt
apscheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
idna==3.4
importlib-metadata==1.7.0
mongoengine==0.24.2
munch==2.5.0
pymongo==4.3.2
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0
requests==2.28.1
setuptools==65.5.1
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
zipp==3.10.0
